requirements.txt
django==1.6.11
jsonschema 
django-extensions==1.6.7
bcrypt
mysql-python==1.2.5
PyJWT==1.5.0